Removes unnessary redundant downloads to speed up re-provisioning. Fixes $PATH so that conversion between hadoop2 and hadoop3 is easier.

+++ vagrant_provision.sh
@@ -15,11 +15,13 @@ apt-get install -y git build-essential cmake automake autoconf libtool libboost-
 
 apt-get install -y yasm
 
-wget --quiet https://github.com/google/protobuf/releases/download/v3.0.0/protobuf-cpp-3.0.0.tar.gz
-tar -xf protobuf-cpp-3.0.0.tar.gz
-rm protobuf-cpp-3.0.0.tar.gz
-cd protobuf-3.0.0; ./autogen.sh && ./configure --prefix=/usr && make && make install
-cd /home/vagrant/; ldconfig
+if [ ! -d /home/vagrant/protobuf-3.0.0 ]; then
+    wget --quiet https://github.com/google/protobuf/releases/download/v3.0.0/protobuf-cpp-3.0.0.tar.gz
+    tar -xf protobuf-cpp-3.0.0.tar.gz
+    rm protobuf-cpp-3.0.0.tar.gz
+    cd protobuf-3.0.0; ./autogen.sh && ./configure --prefix=/usr && make && make install
+    cd /home/vagrant/; ldconfig
+fi
 
 # Install and setup dependencies of hadoop
 apt-get install -y ssh pdsh openjdk-8-jdk-headless
@@ -35,6 +37,9 @@ alias hadoop3="/home/vagrant/hadoop3/bin/hadoop"
 if [ -d /home/vagrant/hadoop3 ]; then
     rm -rf /home/vagrant/hadoop3
 fi
+if [ -f hadoop-3.0.0-beta1-2.tar.gz ]; then
+    rm -f hadoop-3.0.0-beta1-2.tar.gz
+fi
 wget --quiet http://kevinlin.web.rice.edu/static/hadoop-3.0.0-beta1-2.tar.gz
 tar -xf hadoop-3.0.0-beta1-2.tar.gz
 mv hadoop-3.0.0-beta1 /home/vagrant/hadoop3
@@ -47,14 +52,18 @@ cat /home/vagrant/rdfs/config/hdfs-site.xml > /home/vagrant/hadoop3/etc/hadoop/h
 cat /home/vagrant/rdfs/config/core-site.xml > /home/vagrant/hadoop3/etc/hadoop/core-site.xml
 
 # add hadoop to path
-echo 'export PATH=/home/vagrant/hadoop2/bin:$PATH' >> /home/vagrant/.bashrc
+echo 'export PATH=${PATH//:\/home\/vagrant\/hadoop2\/bin:/:}' >> /home/vagrant/.bashrc
+echo 'export PATH=/home/vagrant/hadoop/bin:$PATH' >> /home/vagrant/.bashrc
 
 # add hadoop to classpath
 echo 'export CLASSPATH=/home/vagrant/hadoop3/share/hadoop/hdfs/*:/home/vagrant/hadoop3/share/hadoop/common/*' >> /home/vagrant/.bashrc
 
-# Download hadoop 2.7.4 as well, but do not set as default.
+# Download hadoop 2.8.1 as well
 if [ -d /home/vagrant/hadoop2 ]; then
     rm -rf /home/vagrant/hadoop2
+fi
+if [ -f hadoop-2.8.1.tar.gz ]; then
+    rm -f hadoop-2.8.1.tar.gz
 fi
 wget --quiet http://kevinlin.web.rice.edu/static/hadoop-2.8.1.tar.gz
 tar -xf hadoop-2.8.1.tar.gz
@@ -71,6 +80,9 @@ if [ -d /home/vagrant/isal ]; then
     cd /home/vagrant
     rm -rf /home/vagrant/isal
 fi
+if [ -f isal-2.tar.gz ]; then
+    rm -f isal-2.tar.gz
+fi
 wget --quiet http://kevinlin.web.rice.edu/static/isal-2.tar.gz
 tar -xf isal-2.tar.gz
 rm isal-2.tar.gz
@@ -84,6 +96,9 @@ cd /home/vagrant
 # Setup Apache zookeeper
 if [ -d /home/vagrant/zookeeper ]; then
     rm -rf /home/vagrant/zookeeper
+fi
+if [ -f zookeeper-3.4.9.tar.gz ]; then
+    rm -f zookeeper-3.4.9.tar.gz
 fi
 wget --quiet http://kevinlin.web.rice.edu/static/zookeeper-3.4.9.tar.gz
 tar -xf zookeeper-3.4.9.tar.gz
@@ -143,32 +158,40 @@ apt-get --assume-yes install libtool
 autoreconf -if
 ./configure
 make && make install
+cd /home/vagrant
 
 # Add Google Mock
-apt-get install -y google-mock
-cd /usr/src/gmock
-cmake CMakeLists.txt
-make
-cp *.a /usr/lib
+if [! -d /usr/src/gmock ]; then
+    apt-get install -y google-mock
+    cd /usr/src/gmock
+    cmake CMakeLists.txt
+    make
+    cp *.a /usr/lib
+    cd /home/vagrant
+fi
 
 # Add Google Test
-apt-get install -y libgtest-dev
-cd /usr/src/gtest
-cmake CMakeLists.txt
-make
-cp *.a /usr/lib
-cd /home/vagrant
+if [ ! -d /usr/src/gtest ]; then
+    apt-get install -y libgtest-dev
+    cd /usr/src/gtest
+    cmake CMakeLists.txt
+    make
+    cp *.a /usr/lib
+    cd /home/vagrant
+fi
 
 # Add Valgrind
 sudo apt-get install -y libc6-dbg
-mkdir valgrindtemp
-cd valgrindtemp
-wget --quiet http://valgrind.org/downloads/valgrind-3.11.0.tar.bz2
-tar -xf valgrind-3.11.0.tar.bz2
-cd valgrind-3.11.0
-./configure --prefix=/usr && sudo make && sudo make install
-cd ../..
-rm -r valgrindtemp
+if [ ! -d /usr/lib/valgrind ]; then
+    mkdir valgrindtemp
+    cd valgrindtemp
+    wget --quiet http://valgrind.org/downloads/valgrind-3.11.0.tar.bz2
+    tar -xf valgrind-3.11.0.tar.bz2
+    cd valgrind-3.11.0
+    ./configure --prefix=/usr && sudo make && sudo make install
+    cd ../..
+    rm -r valgrindtemp
+fi
 
 # Download demo tables.
 if [ ! -d /home/vagrant/demo_script ]; then
